[Lesson06] Modification in calculating the stagnation points.
If the circulation strength is low enough, the two stagnation points are on the bottom half of the cylinder.
If the circulation strength is high, the two stagnation points are the vertical axis, one inside and the other outside the cylinder.

Add a function get_stagnation_points() that computes the location of the stangation points depending on the circulation strength of the vortex.

Add some text to explain, how the location of the stagnation points is calculated.