Make new branch for second order solver
Added linearly varying vortex panels. This is a fairly small change
to Panel. Mostly I've added the additional terms in the velocity
function and then call them.

New routine solve_gamma_O2 (which uses _construct_A_b_O2) uses
this to get a much more accurate solution, but it still has the
same arguments and still saves the panel center strength as gamma.

Hopefully, this doesn't change ANY previous workflow or results,
but enables O(2) simulations if desired.

Changed 3_6 to include this feature, but I'm not sure that is the
best place yet.