Simplify definition of figure sizes + other things
Simplify the way to set up axis limits (previous way was cumbersome for what we
are doing). Now we use `pyplot.gca().set_aspect()` to automatically adjust the
size of the figure.

Correct some small typos.

Add missing label `theoretical` to the figure legend.

Use `.format()` for `print` statements.

Simplify the way to create the mesh in the domain. The old way was cumbersome;
now we specify explicitly the dimensions of the domain.