Fix bug in function and simplify axis setup
Simplify the way to set axis limits (set the limits explicitly while keeping the
aspect equal).

Correct the function `define_panels`. We were not creating the Numpy array
`panels` and the code could not run... How did we miss that?!

Correct small typo.

Calculate the chord before the lift coefficient (have to do this after we
simplified the way to set axis limits).