Use `numpy.vectorize` and make figures prettier
Lesson 08:

* Remove `contourf` plots to display the stagnation points. Instead, we compute
the indices of the point and use `scatter` to display it on the figure.

* Add an example of using the `lambda` function with multiple parameters (the
first one being the variable of integration).

* Use `numpy.vectorize()` to avoid the double loop over the computational domain
(I found that to be computationally more efficient in our case). Also, we add an
explanation about this new function.

Lesson 09:

* Remove useless `lambda` functions.

* Use argument `label` directly in `pyplot.plot()` to tag the plot, instead of
doing it in `pyplot.legend()` (basically moving the plot tag closer to the plot
definition).

Lesson 10:

* Use `numpy.vectorize()` and add a reference to Lesson 08 where we used the
function for the first time.

Lesson 11:

* Remove useless `lambda` function.