Update boundary layer integration methods
I changed the internals of march to integrate for the momentum
thickness (squared) using an implicit trapz method. This is
more accurate and cleaner to code.

I also added an implementation of Thwaites approximation. This
is less accurate (especially near the leading edge) but avoids
the implicit step to find separation.

Tested on flat plate and circle.