Clean-up influence functions
The matrix construction routines relied on overloading the values
of gamma in the velocity function. This required 6 influence function
evaluations for each O2 matrix coefficient, making O2 6x slower.

Now there are _constant and _linear Panel influence functions and
these are called for the construction and the velocity evaluations.
This cleans up the code and makes O2 less than 2x slower than O1.